// src/constants/actions/action.enum.ts
// Define Action Types - Using const enums for better type safety
export const enum ActionType {
    SET_USER = "SET_USER",
    SET_USER_ERROR = "SET_USER_ERROR",
    SET_USER_LOADING = "SET_USER_LOADING",
    SET_QUESTIONS = "SET_QUESTIONS",
    SET_CURRENT_QUESTION_INDEX = "SET_CURRENT_QUESTION_INDEX",
    SET_QUESTIONS_ERROR = "SET_QUESTIONS_ERROR",
    SET_QUESTIONS_LOADING = "SET_QUESTIONS_LOADING",
    TOGGLE_MODAL = "TOGGLE_MODAL",
    SET_COMPLETE = "SET_COMPLETE",
  }